move standford to dependency dir
