--- src/cltkv1/wrappers/__init__.py
@@ -1,1 +1,0 @@
-"""Init for `cltkv1.wrappers`."""
